requirements.txt
click==7.1.2
dnspython==1.16.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymongo==3.11.2
Werkzeug==1.0.1
gunicorn==20.0.4
python-dotenv==0.17.1